* Phase7Sprint8: (29 commits)
  Fixed bug#2779 - A, EST & ESTA Class extensions to be included in the Labour Rate rules.
  SQL for bug#2754 (Production - remove liability update in Awaiting Invoice Payment) updated with db clean script.
  Fixed bug#2755 7.8.6 - Incorrect BRE Failure Messages for 'CHO's Client VAT Registered Check'.
  Minor tool-tip update.
  Added sql for bug#2754 - Production - remove liability update in Awaiting Invoice Payment.
  Fixed minor bugs.
  Corrected attachment file upload size : increased size allowed in upload by struts interceptor.
  Completed todo item 7.8.6 - New Business Rules.
  Removed option from batch update menu to move claims to InvoicePaymentLogged for Manual claims.
  Remaining business rules added for todo item 7.8.6.
  Corrected upload limit & message.
  Corrected upload limit to 20mb.
  First three new BRE rules implemented for todo item 7.8.6 (New Business Rules) and ui/db adjustments implemented for 3 remaining rules. 3 remaining rules still to be implemented.
  Removed unwanted items from more-actions drop-down for insurer claims at new statuse : 'Final Review' and 'Make Interim Payment'.
  Completed updates to todo item 7.8.5 Report and Dashboard updates.
  Added new rows requested.
  Implemen ted todo item 7.8.5 Report & Dashboard updates required for change in 'Insurer Upload' workflow.
  Added null pointer protection and removed unused local variables.
  BUG#2726 - Production - Claim Type Filter not active after batch update
  BUG#2726 - Production - Claim Type Filter not active after batch update
  ...

Conflicts:
	idas.chox.web/src/main/resources/logback-localServer.xml

+++ idas.chox.web/src/main/java/idas/chox/web/actions/ClaimAction.java
@@ -1167,11 +1167,17 @@ public class ClaimAction extends BaseAction implements ModelDriven<Claim>, Prepa
                 }
             
                 // Remove 'Mark Claim For Supplementary Invoice(s)' for Insure (Manual) invoices (bug#2586)
+                 // The following should really be done by updating the accessibility tables....
                 else if (actionName.equals(ExtraAction.MARK_SUPPLEMENTARY_INVOICED_CLAIM)
                         && claim.getClaimType() == ClaimType.INSURER_INVOICE) {
                     accessRight = 0;
                 }
-                
+                 // The following should really be done by updating the accessibility tables....
+                 else if ((actionName.equals(ExtraAction.MAKE_INTERIM_PAYMENT) || actionName.equals(ExtraAction.FINAL_REVIEW))
+                         && ClaimType.isInsurerUpload(claim.getClaimType())) {
+                     accessRight = 0;
+                 }
+
                 // bug#2719 - disable update of workgrouup/owner if not already routed/assigned
                 else if (actionName.equals(ExtraAction.UPDATE_CLAIM_WORKGROUP_AND_OWNER)
                         && (claim.getWorkgroup() == null || claim.getClaimOwner() == null)) {
@@ -1362,9 +1368,9 @@ public class ClaimAction extends BaseAction implements ModelDriven<Claim>, Prepa
 
     public String removeNotification() {
 
-        if (notificationId > 0) {
+        if (notificationId != null && notificationId > 0) {
             notificationService.removeNotificationById(notificationId);
-        } else if (notificationService != null && claim != null) { // No id given so remove all notifications
+        } else if (notificationId != null && notificationService != null && claim != null) { // No id given so remove all notifications
             if (getIsInsurer()) {
                 notificationService.removeAllInsurerNotifications(claim.getId());
             } else {
@@ -1377,9 +1383,9 @@ public class ClaimAction extends BaseAction implements ModelDriven<Claim>, Prepa
 
     public String acknowledgeNotification() {
 
-        if (notificationId > 0) {
+        if (notificationId != null && notificationId > 0) {
                 notificationService.acknowledgeNotificationById(notificationId);
-        } else if (notificationService != null && claim != null) {
+        } else if (notificationId != null && notificationService != null && claim != null) {
             LOG.debug("Acknowledge All Notifications");
             if (getIsInsurer()) {
                 LOG.debug("Acknowledge All Notifications for Insurer ");
@@ -1462,6 +1468,14 @@ public class ClaimAction extends BaseAction implements ModelDriven<Claim>, Prepa
 
     public boolean getIsInsurerInvoice() {
         return claim.getClaimType() == ClaimType.INSURER_INVOICE;
+    }
+
+    public boolean getIsInsurerManual() {
+        return claim.getClaimType() == ClaimType.INSURER_INVOICE
+                || claim.getClaimType() == ClaimType.INSURER_CLAIM
+                || claim.getClaimType() == ClaimType.INSURER_ORIGINAL_INVOICE
+                || claim.getClaimType() == ClaimType.INSURER_SUPPLEMENTARY_INVOICE
+                || claim.getClaimType() == ClaimType.INSURER_UPLOAD;
     }
 
     public void setIntelligentNoteDisplayEngine(IntelligentNoteDisplayEngine intelligentNoteDisplayEngine) {
